type builds and scrol lto top bug fix
<!-- ps-id: 8167a06f-08c0-4831-a8f6-6d41c3515841 -->
