+++ apps/client/src/app.tsx
@@ -3,7 +3,6 @@ import Router from './router';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { loadFromStorage } from '@repo/auth';
 import { Toaster } from '@/components/toast/toaster';
-import { useLocation } from 'react-router-dom';
 
 const queryClient = new QueryClient();
 
@@ -24,7 +23,6 @@ function App() {
   return (
     <>
       <QueryClientProvider client={queryClient}>
-        <ScrollToTop />
         <Router />
         <Toaster />
       </QueryClientProvider>
@@ -33,13 +31,3 @@ function App() {
 }
 
 export default App;
-
-function ScrollToTop() {
-  const { pathname } = useLocation();
-
-  useEffect(() => {
-    window.scrollTo(0, 0);
-  }, [pathname]);
-
-  return null;
-}
